Create hyperpoetry
This makes a link to the hyperpoetry, which contains a link to itself, which contains a link to itself, which contains a link to itself, all of which are different forms from the previous, but are seemingly identical. Possible work in the future to create a true regression that is self-referential.